requirements.txt
cv2
numpy~=1.19.5
opencv-python~=4.5.3.56
matplotlib~=3.1.2
pandas~=1.2.5
rospy~=1.15.11
argparse~=1.4.0
Pillow~=8.3.1
imutils~=0.5.4
tensorflow~=2.5.0
onnxruntime~=1.8.1
onnx~=1.9.0